# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Marketplace Nigeria API Documentation'
  description: 'Complete REST API documentation for Marketplace Nigeria platform. This API allows you to manage adverts, users, messages, payments, and more.'
  version: '3.0'
servers:
  -
    url: 'http://127.0.0.1:8030'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can obtain your authentication token by logging in through the `/api/login` endpoint. The token should be included in the Authorization header as a Bearer token for all authenticated requests.'
security:
  -
    default: []
paths:
...
  /api/payments/initialize:
    post:
      summary: ''
      operationId: postApiPaymentsInitialize
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipping_data:
                  type: object
                  description: ''
                  example: []
                  properties:
                    first_name:
                      type: string
                      description: ''
                      example: aut
                    last_name:
                      type: string
                      description: ''
                      example: laudantium
                    phone:
                      type: string
                      description: ''
                      example: nemo
                    grand_total:
                      type: number
                      description: ''
                      example: 8.0
                    commission:
                      type: number
                      description: ''
                      example: 46357.5363968
                    shipping_cost:
                      type: number
                      description: ''
                      example: 553583.5193
                    shipping_method:
                      type: integer
                      description: ''
                      example: 3
                    ad:
                      type: object
                      description: ''
                      example: []
                      properties:
                        id:
                          type: integer
                          description: 'The <code>id</code> of an existing record in the adverts table.'
                          example: 18
                        price:
                          type: number
                          description: ''
                          example: 166.97726
                      required:
                        - id
                        - price
                    reciever_city:
                      type: object
                      description: ''
                      example: []
                      properties:
                        id:
                          type: integer
                          description: ''
                          example: 8
                          nullable: true
                    reciever_state:
                      type: object
                      description: ''
                      example: []
                      properties:
                        id:
                          type: integer
                          description: ''
                          example: 15
                          nullable: true
                  required:
                    - first_name
                    - last_name
                    - phone
                    - grand_total
                    - commission
                    - shipping_cost
                    - shipping_method
              required:
                - shipping_data
      security: []
  /api/payments/initialize-boost:
    post:
      summary: ''
      operationId: postApiPaymentsInitializeBoost
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advert_id:
                  type: integer
                  description: 'The <code>id</code> of an existing record in the adverts table.'
                  example: 7
                amount:
                  type: number
                  description: 'Must be at least 0.'
                  example: 15
                boost_type:
                  type: string
                  description: ''
                  example: labore
                duration:
                  type: integer
                  description: 'Must be at least 1.'
                  example: 77
              required:
                - advert_id
                - amount
                - boost_type
                - duration
      security: []
  '/api/payments/{paymentId}':
    get:
      summary: ''
      operationId: getApiPaymentsPaymentId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Payments
      security: []
    parameters:
      -
        in: path
        name: paymentId
        description: ''
        example: est
        required: true
        schema:
          type: string
  '/api/payments/user/{userId}':
    get:
      summary: ''
      operationId: getApiPaymentsUserUserId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Payments
      security: []
    parameters:
      -
        in: path
        name: userId
        description: ''
        example: dolores
        required: true
        schema:
          type: string
  '/api/boosts/user/{userId}':
    get:
      summary: ''
      operationId: getApiBoostsUserUserId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Payments
      security: []
    parameters:
      -
        in: path
        name: userId
        description: ''
        example: cum
        required: true
        schema:
          type: string
  /api/payments/callback:
    post:
      summary: ''
      operationId: postApiPaymentsCallback
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reference:
                  type: string
                  description: ''
                  example: et
              required:
                - reference
      security: []
...